Add files via upload
You will be writing a program that runs an ATM (Automated Teller Machine), also called a "cash machine". The program will display options on the screen and the user will select an option by typing on the keyboard. In real life, the user would press a touch screen, instead. We will assume that the ATM is the only interaction that the user has with their bank account. There is no physical bank at which your customer can make changes to their accounts, outside of your program.